django/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via
    #   -r requirements.in
    #   kombu
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   django
baker==1.3
    # via -r requirements.in
billiard==4.2.1
    # via
    #   -r requirements.in
    #   celery
black==24.10.0
    # via -r requirements.in
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
celery==5.4.0
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   -r requirements.in
    #   requests
cffi==1.17.1
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==3.4.0
    # via
    #   -r requirements.in
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.1
    # via
    #   -r requirements.in
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements.in
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements.in
    #   celery
coverage==7.6.4
    # via -r requirements.in
cryptography==43.0.3
    # via
    #   -r requirements.in
    #   social-auth-core
defusedxml==0.8.0rc2
    # via
    #   -r requirements.in
    #   python3-openid
    #   social-auth-core
django==5.1.3
    # via
    #   -r requirements.in
    #   django-filter
    #   django-redis
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-yasg
    #   model-bakery
    #   social-auth-app-django
django-filter==24.3
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-stubs-ext==5.1.1
    # via -r requirements.in
django-templated-mail==1.1.1
    # via
    #   -r requirements.in
    #   djoser
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via
    #   -r requirements.in
    #   djoser
djoser==2.2.3
    # via -r requirements.in
drf-yasg==1.21.8
    # via -r requirements.in
flake8==7.1.1
    # via -r requirements.in
idna==3.10
    # via
    #   -r requirements.in
    #   requests
inflection==0.5.1
    # via
    #   -r requirements.in
    #   drf-yasg
isort==5.13.2
    # via -r requirements.in
kombu==5.4.2
    # via
    #   -r requirements.in
    #   celery
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
model-bakery==1.20.0
    # via -r requirements.in
mypy==1.13.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
    #   mypy
newrelic==10.2.0
    # via -r requirements.in
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
packaging==24.1
    # via
    #   -r requirements.in
    #   black
    #   build
    #   drf-yasg
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   black
pillow==11.0.0
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via
    #   -r requirements.in
    #   black
prompt-toolkit==3.0.48
    # via
    #   -r requirements.in
    #   click-repl
psycopg2-binary==2.9.10
    # via
    #   -r requirements.in
    #   psycopg2-pool
psycopg2-pool==1.2
    # via -r requirements.in
pycodestyle==2.12.1
    # via
    #   -r requirements.in
    #   flake8
pycparser==2.22
    # via
    #   -r requirements.in
    #   cffi
pyflakes==3.2.0
    # via
    #   -r requirements.in
    #   flake8
pyjwt==2.9.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   social-auth-core
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   celery
python-decouple==3.8
    # via -r requirements.in
python3-openid==3.2.0
    # via
    #   -r requirements.in
    #   social-auth-core
pytz==2024.2
    # via
    #   -r requirements.in
    #   drf-yasg
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   drf-yasg
redis==5.2.0
    # via
    #   -r requirements.in
    #   django-redis
requests==2.32.3
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via
    #   -r requirements.in
    #   social-auth-core
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
social-auth-app-django==5.4.2
    # via
    #   -r requirements.in
    #   djoser
social-auth-core==4.5.4
    # via
    #   -r requirements.in
    #   social-auth-app-django
sqlparse==0.5.1
    # via
    #   -r requirements.in
    #   django
types-pyyaml==6.0.12.20240917
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   django-stubs-ext
    #   mypy
tzdata==2024.2
    # via
    #   -r requirements.in
    #   celery
    #   kombu
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   drf-yasg
urllib3==2.2.3
    # via
    #   -r requirements.in
    #   requests
vine==5.1.0
    # via
    #   -r requirements.in
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements.in
    #   prompt-toolkit
wheel==0.44.0
    # via
    #   -r requirements.in
    #   pip-tools
whitenoise==6.8.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
